autorl_project/requirements.txt
Appium-Python-Client
selenium
psutil
prometheus_client
Pillow
flask
flask-cors


pytesseract
